Test: rename cronosd to tgenesisd
